+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 * CXV1-4415 - Campaign settings UI improvements
+* CXV1-4676 - Campaign settings bug fix
 
 ## [4.1.17]
 * CXV1-4390 - Got campaigns properly enabling and disabling